pyproject.toml
[project]
name = "tongui-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.6.0",
    "flash-attn>=2.7.4.post1",
    "gradio>=5.25.2",
    "numpy>=2.2.4",
    "pillow>=11.2.1",
    "transformers",
    "vllm>=0.8.4",
]

[tool.uv.sources]
transformers = { git = "https://github.com/huggingface/transformers.git" }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["tongui"]

